% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweetcleanerfunction.R
\name{tweetcleaner}
\alias{tweetcleaner}
\title{Tweet Cleaner}
\usage{
tweetcleaner(text)
}
\arguments{
\item{text}{Any text/ character string of tweets}
}
\description{
This function allows users to clean tweets for sentiment analysis
}
\examples{
# example 
tweetcleaner()
}
\keyword{text}
